10/05/2023 Modified: 1.Fixed Serializers to allow blank/null 2. Modified Models

+++ digital_dex_admin_web/models/exempt_assessment_model.py
@@ -20,14 +20,14 @@ class ExemptAssessment(models.Model):
         ('I', 'Industrial'),
         ('S', 'Special'),
         ('T', 'Timberland'),
-        ('M', 'Mineral'),   
+        ('M', 'Mineral'),
     )
     classification = models.CharField(max_length=1, choices=CLASSIFICATION_CHOICES, null = True)
-    assessed_value = models.IntegerField(null = True)
+    assessed_value = models.FloatField(null = True)
     legal_basis = models.CharField(max_length=255, null = True)
     effectivity = models.CharField(max_length=255, null = True)
     remarks = models.CharField(max_length=255, null = True)
-    
+
 
     class Meta:
             verbose_name = 'Exempt Assessment'
