+++ digital_dex_admin_web/models/taxable_assessment_model.py
@@ -20,15 +20,15 @@ class TaxableAssessment(models.Model):
         ('I', 'Industrial'),
         ('S', 'Special'),
         ('T', 'Timberland'),
-        ('M', 'Mineral'),   
+        ('M', 'Mineral'),
     )
     classification = models.CharField(max_length=1, choices=CLASSIFICATION_CHOICES, null = True)
-    assessed_value = models.IntegerField(null = True)
+    assessed_value = models.FloatField(null = True)
     prev_arpn = models.CharField(max_length=255, null = True)
     prev_td_no = models.CharField(max_length=255, null = True)
     effectivity = models.CharField(max_length=255, null = True)
     remarks = models.CharField(max_length=255, null = True)
-    
+
 
     class Meta:
             verbose_name = 'Taxable Assessment'
